feat: Community 코디네이터 추가
